+++ UI/src/com/kotcrab/vis/ui/widget/file/internal/FileChooserText.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.kotcrab.vis.ui.widget.file;
+package com.kotcrab.vis.ui.widget.file.internal;
 
 import com.badlogic.gdx.utils.I18NBundle;
 import com.kotcrab.vis.ui.VisUI;